+++ pel216/week6/inc/integration/RectangleRule.h
@@ -31,7 +31,7 @@ namespace pel216 {
 			/**
 			 * @see pel216::week6::IntegrationRule#evaluate()
 			 */
-			virtual double doEvaluate(Function *function, double a, double b, size_t intervals) {
+			virtual double doEvaluate(Function *function, double a, double b, double intervals) {
 
 				double result = 0.0f;
 
@@ -39,7 +39,7 @@ namespace pel216 {
 
 				double a_prime = a;
 				double b_prime = a_prime + alpha;
-				for (size_t idx = 0; idx < intervals; idx++) {
+				for (double idx = 0.0f; idx < intervals; idx++) {
 					double a_plus_b_div_2 = (a_prime + b_prime) / 2;
 					result += (b_prime - a_prime) * function->evaluate(a_plus_b_div_2);
 					a_prime += alpha;
